requirements.txt
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
Faker==9.5.2
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.26
text-unidecode==1.3
Werkzeug==2.0.2
